+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## 0.3.3
+- Fixed missing options param
+
 ## 0.3.2
 - Added `RestrrBuilder#refresh`
 - Fixed `RestrrOptions` and added `isWeb` attribute